tests/requirements-test.txt
# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0  # Parallel test execution
pytest-html>=4.1.0  # HTML reports
pytest-sugar>=0.9.7  # Better output

# Testcontainers for Integration Tests
testcontainers[postgres,kafka,redis]>=3.7.0

# HTTP Testing
httpx>=0.25.0  # Async HTTP client
requests>=2.31.0
responses>=0.24.0  # Mock HTTP responses

# Mocking
pytest-mock>=3.12.0
unittest-mock>=1.4.0
factory-boy>=3.3.0  # Test data factories
faker>=20.0.0  # Fake data generation

# Async Testing
aioresponses>=0.7.4
asynctest>=0.13.0

# Kafka Testing
aiokafka>=0.9.0
kafka-python>=2.0.2

# Database Testing
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Redis Testing
redis>=5.0.0

# Load Testing
locust>=2.20.0

# Security Testing
pip-audit>=2.6.1
bandit>=1.7.6
safety>=2.3.5

# Code Quality
mypy>=1.7.0
ruff>=0.1.6
black>=23.11.0

# Snapshot Testing
syrupy>=4.6.0

# Contract Testing
pact-python>=2.2.0
schemathesis>=3.21.0

# Performance Profiling
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0
